requirements.txt
streamlit==1.40.1
anthropic==0.84.0
edge-tts==7.2.7
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
aiohttp==3.10.11
